datahike now supports connections to postgresql via ssl encryption
Since PostgreSQL supports encrypted connections there is the
need to add this option to datahike as well. I added the
corresponding configuration and adapted the datahike-postgres
library to the new options and the new configuration format.

It seemed to make more sense to switch datahike-postgres
to pass the configuration as map because konserve-pg is able
to handle a map anyways. Converting a string to map to string
did not seem reasonable so I switched the lib to use map as
well without breaking the api.

- PR #6 in replikative/datahike-postgres
- Closes #133