Apply CircleCI Orb to normalize CI
With a centralized Orb it is possible to make it easier for people
to implement CI themselves based on the examples. One only has to
pick the necessary steps and it is good to go. Orbs do not help
with a centralized workflow because there is no way to set a
default workflow within an Orb so steps still have to be inserted
manually into all the projects. Also the Orb version has to be
manually adjusted in case there is a newer release but this is
now set to use the latest major version, currently '0'.

- Refers #167
- Renaming of scripts to run tests because of bad English