Switch to Clojure CLI tools
This commit switches over the CI from leiningen to clojure cli tools.
For that it introduces the file deps.edn and pom.xml that are needed
for clojure cli tools to resolve dependencies and build the artifact.
The project.clj is deleted.

The CI is done by CircleCI and the old circle config is delete, but
the config.yml in .circleci is updated to use the tools orb. The
setup step is taken out of the orb and customized to use maven to
compile the java source code. That step was not clear in the
beginning and took me a while to figure out.

- Introduce deps.edn and pom.xml
- Delete project.clj
- Update CI
- Leave out all CLJS functionality for now
- Adapt documentation
- Closes #253
- With contributions by elMor3no and groundedSAGE