Move to main branch (#454)
* Switch to GitHub Flow and tools.build

This commit changes the CI to create a release on every
merge to the `main`-branch. The patch-release will always
incremented when released. CircleCI-pipeline creates
a release automatically on GitHub and tags the commit.

This commit also switches to use tools.build and with
that it is possible to run the compilation and jar-builds
from the REPL. There is no need to run maven anymore.

* Documentation and ADR

Co-authored-by: Timo Kramer <fw-git@timokramer.de>
Co-authored-by: Timo Kramer <info@lambdaforge.io>